requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
websockets==12.0
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.21.1
